+++ src/systems/__tests__/InitSpaces.test.ts
@@ -1,7 +1,8 @@
 import { expect, describe, test } from 'bun:test';
 import { Core } from '../../core';
-import { EntityAdapter, PlayerEntityState, SpaceState } from '../EntityAdapter';
+import { EntityAdapter } from '../EntityAdapter';
 import { InitSpaces } from '../InitSpaces';
+import { PlayerEntityState, SpaceState } from '../../components';
 
 describe('InitSpaces', () => {
 	const core = new Core();
